+++ aiku/aiku-main/src/main/java/aiku_main/repository/betting/BettingRepositoryCustom.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.betting;
 
 import aiku_main.repository.dto.TeamBettingResultMemberDto;
 import common.domain.Betting;
