+++ aiku/aiku-main/src/main/java/aiku_main/repository/event/EventRepository.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.event;
 
 import common.domain.event.CommonEvent;
 import common.domain.event.RecommendEvent;
