+++ aiku/aiku-main/src/main/java/aiku_main/repository/log/PointLogRepository.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.log;
 
 import common.domain.log.PointLog;
 import org.springframework.data.jpa.repository.JpaRepository;
