+++ aiku/aiku-main/src/main/java/aiku_main/repository/racing/RacingRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.racing;
 
 import aiku_main.dto.MemberProfileResDto;
 import aiku_main.repository.dto.TeamRacingResultMemberDto;
