+++ aiku/aiku-main/src/main/java/aiku_main/repository/team/TeamRepository.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.team;
 
 import common.domain.Status;
 import common.domain.team.Team;
