+++ aiku/aiku-main/src/main/java/aiku_main/repository/title/TitleRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package aiku_main.repository;
+package aiku_main.repository.title;
 
 import aiku_main.dto.member.TitleMemberResDto;
 import com.querydsl.core.types.Projections;
