+++ aiku/aiku-main/src/main/java/aiku_main/service/BettingService.java
@@ -7,9 +7,9 @@ import aiku_main.application_event.publisher.PointChangeEventPublisher;
 import aiku_main.dto.betting.BettingAddDto;
 import aiku_main.exception.BettingException;
 import aiku_main.exception.ScheduleException;
-import aiku_main.repository.BettingRepository;
-import aiku_main.repository.MemberRepository;
-import aiku_main.repository.ScheduleRepository;
+import aiku_main.repository.betting.BettingRepository;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.schedule.ScheduleRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.Betting;
