+++ aiku/aiku-main/src/main/java/aiku_main/service/MemberPointChangeFacade.java
@@ -3,13 +3,7 @@ package aiku_main.service;
 import aiku_main.application_event.event.PointChangeReason;
 import aiku_main.application_event.event.PointChangeType;
 import aiku_main.application_event.publisher.PointChangeFailEventPublisher;
-import aiku_main.exception.MemberNotFoundException;
 import aiku_main.exception.PointChangeFailException;
-import aiku_main.repository.MemberRepository;
-import common.domain.log.PointLog;
-import common.domain.member.Member;
-import common.domain.value_reference.MemberValue;
-import common.exception.NotEnoughPoint;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
