+++ aiku/aiku-main/src/main/java/aiku_main/service/MemberRegisterService.java
@@ -8,8 +8,8 @@ import aiku_main.dto.member.MemberRegisterDto;
 import aiku_main.dto.member.NicknameExistResDto;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.oauth.KakaoOauthHelper;
-import aiku_main.repository.EventRepository;
-import aiku_main.repository.MemberRepository;
+import aiku_main.repository.event.EventRepository;
+import aiku_main.repository.member.MemberRepository;
 import aiku_main.s3.S3ImageProvider;
 import common.domain.event.RecommendEvent;
 import common.domain.member.Member;
