+++ aiku/aiku-main/src/main/java/aiku_main/service/MemberService.java
@@ -5,8 +5,8 @@ import aiku_main.dto.*;
 import aiku_main.dto.member.*;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.exception.TitleException;
-import aiku_main.repository.MemberRepository;
-import aiku_main.repository.TitleRepository;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.title.TitleRepository;
 import aiku_main.s3.S3ImageProvider;
 import common.domain.ServiceAgreement;
 import common.domain.member.Member;
