+++ aiku/aiku-main/src/main/java/aiku_main/service/PointChangeFailSagaService.java
@@ -4,9 +4,8 @@ import aiku_main.application_event.event.PointChangeReason;
 import aiku_main.application_event.event.PointChangeType;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.kafka.KafkaProducerService;
-import aiku_main.repository.MemberRepository;
+import aiku_main.repository.member.MemberRepository;
 import aiku_main.service.strategy.RollbackProcessor;
-import common.domain.value_reference.MemberValue;
 import common.kafka_message.KafkaTopic;
 import common.kafka_message.alarm.AlarmMessageType;
 import common.kafka_message.alarm.PointErrorMessage;
