+++ aiku/aiku-main/src/main/java/aiku_main/service/RacingService.java
@@ -4,8 +4,8 @@ import aiku_main.dto.racing.ScheduleRacing;
 import aiku_main.dto.racing.ScheduleRacingMember;
 import aiku_main.dto.racing.ScheduleRacingResult;
 import aiku_main.exception.BettingException;
-import aiku_main.repository.RacingRepository;
-import aiku_main.repository.ScheduleRepository;
+import aiku_main.repository.racing.RacingRepository;
+import aiku_main.repository.schedule.ScheduleRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.Racing;
