+++ aiku/aiku-main/src/main/java/aiku_main/service/ScheduleService.java
@@ -9,9 +9,9 @@ import aiku_main.dto.schedule.*;
 import aiku_main.exception.ScheduleException;
 import aiku_main.exception.TeamException;
 import aiku_main.kafka.KafkaProducerService;
-import aiku_main.repository.MemberRepository;
-import aiku_main.repository.ScheduleRepository;
-import aiku_main.repository.TeamRepository;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.schedule.ScheduleRepository;
+import aiku_main.repository.team.TeamRepository;
 import aiku_main.scheduler.ScheduleScheduler;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
