+++ aiku/aiku-main/src/main/java/aiku_main/service/TeamService.java
@@ -5,9 +5,13 @@ import aiku_main.application_event.publisher.TeamEventPublisher;
 import aiku_main.dto.*;
 import aiku_main.exception.MemberNotFoundException;
 import aiku_main.exception.TeamException;
-import aiku_main.repository.*;
+import aiku_main.repository.betting.BettingRepository;
 import aiku_main.repository.dto.TeamBettingResultMemberDto;
 import aiku_main.repository.dto.TeamRacingResultMemberDto;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.racing.RacingRepository;
+import aiku_main.repository.schedule.ScheduleRepository;
+import aiku_main.repository.team.TeamRepository;
 import common.util.ObjectMapperUtil;
 import common.domain.member.Member;
 import common.domain.schedule.Schedule;
