+++ aiku/aiku-main/src/test/java/aiku_main/service/MemberFirebaseServiceTest.java
@@ -2,7 +2,7 @@ package aiku_main.service;
 
 import aiku_main.dto.FirebaseTokenDto;
 import aiku_main.exception.MemberNotFoundException;
-import aiku_main.repository.MemberRepository;
+import aiku_main.repository.member.MemberRepository;
 import common.domain.member.Member;
 import common.domain.member.MemberProfileBackground;
 import common.domain.member.MemberProfileCharacter;
