+++ aiku/aiku-main/src/test/java/aiku_main/service/PointChangeEventTest.java
@@ -4,7 +4,7 @@ import aiku_main.application_event.event.PointChangeEvent;
 import aiku_main.application_event.event.PointChangeReason;
 import aiku_main.application_event.event.PointChangeType;
 import aiku_main.application_event.handler.PointChangeEventHandler;
-import aiku_main.repository.PointLogRepository;
+import aiku_main.repository.log.PointLogRepository;
 import common.domain.log.EventLog;
 import common.domain.log.PointLog;
 import common.domain.log.PointLogStatus;
@@ -12,7 +12,6 @@ import common.domain.member.Member;
 import common.domain.member.MemberProfileBackground;
 import common.domain.member.MemberProfileCharacter;
 import common.domain.member.MemberProfileType;
-import common.domain.value_reference.MemberValue;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
