+++ aiku/aiku-main/src/test/java/aiku_main/service/RacingServiceTest.java
@@ -3,7 +3,7 @@ package aiku_main.service;
 import aiku_main.dto.racing.ScheduleRacing;
 import aiku_main.dto.racing.ScheduleRacingMember;
 import aiku_main.dto.racing.ScheduleRacingResult;
-import aiku_main.repository.RacingRepository;
+import aiku_main.repository.racing.RacingRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.Location;
