+++ aiku/aiku-main/src/test/java/aiku_main/service/ScheduleServiceTest.java
@@ -6,9 +6,9 @@ import aiku_main.dto.*;
 import aiku_main.dto.schedule.*;
 import aiku_main.exception.ScheduleException;
 import aiku_main.exception.TeamException;
-import aiku_main.repository.MemberRepository;
-import aiku_main.repository.ScheduleRepository;
-import aiku_main.repository.TeamRepository;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.schedule.ScheduleRepository;
+import aiku_main.repository.team.TeamRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.*;
