+++ aiku/aiku-main/src/test/java/aiku_main/service/TeamServiceIntegrationTest.java
@@ -3,8 +3,8 @@ package aiku_main.service;
 import aiku_main.dto.team.*;
 import aiku_main.dto.team.TeamResDto;
 import aiku_main.exception.TeamException;
-import aiku_main.repository.MemberRepository;
-import aiku_main.repository.TeamRepository;
+import aiku_main.repository.member.MemberRepository;
+import aiku_main.repository.team.TeamRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.*;
